Make the rollover tests actually roll over
ns_alarm_state and ns_melody typed milliseconds as unsigned long. That is
32-bit on the ESP32 but 64-bit on the native host, so 0xFFFFFFFF plus a
few minutes never wrapped and the 49-day guarantee was never exercised.
Written to wrap for real, the tests failed on the host: a wrap mid-melody
stalled playback and a snooze taken just before a wrap read as expired.

Both libraries now use uint32_t for every millisecond value, matching
millis(), and the melody sequencer compares due times as int32_t. The
tests assert that they wrap.

Closes #49

+++ cores3/include/alerts.h
@@ -13,8 +13,8 @@
 struct AlarmState {
     AlarmSchedule sched;
 
-    void snooze(unsigned long nowMs, int level, int timeout_min);
-    unsigned long snoozeRemaining(unsigned long nowMs) const;
+    void snooze(uint32_t nowMs, int level, int timeout_min);
+    uint32_t snoozeRemaining(uint32_t nowMs) const;
 };
 
 /**
